+++ src/pytest_dbt_core/fixtures.py
@@ -8,6 +8,7 @@ import os
 import dbt.tracking
 import pytest
 from _pytest.fixtures import SubRequest
+from dbt import flags
 from dbt.clients.jinja import MacroGenerator
 from dbt.config.runtime import RuntimeConfig
 from dbt.context import providers
@@ -41,6 +42,7 @@ class Args:
 
     project_dir: str
     target: str | None
+    profile: str | None
 
 
 @pytest.fixture
@@ -58,9 +60,15 @@ def config(request: SubRequest) -> RuntimeConfig:
     RuntimeConfig
         The runtime config.
     """
+
+    profiles_dir = request.config.getoption("--profiles-dir")
+    if profiles_dir:
+        flags.PROFILES_DIR = os.path.abspath(profiles_dir)
+
     args = Args(
         project_dir=request.config.getoption("--dbt-project-dir"),
         target=request.config.getoption("--dbt-target"),
+        profile=None,
     )
     config = RuntimeConfig.from_args(args)
     return config